+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/metadata/expressions/CollateFunctionKeyExpression.java
@@ -171,6 +171,11 @@ public class CollateFunctionKeyExpression extends FunctionKeyExpression implemen
         return 1;
     }
 
+    @Override
+    public boolean isFixedColumnSize() {
+        return getArguments().isFixedColumnSize();
+    }
+
     @Nullable
     @Override
     public Function<Object, Object> getComparandConversionFunction() {