+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/metadata/expressions/SplitKeyExpression.java
@@ -102,6 +102,11 @@ public class SplitKeyExpression extends BaseKeyExpression implements AtomKeyExpr
         return splitSize;
     }
 
+    @Override
+    public boolean isFixedColumnSize() {
+        return getJoined().isFixedColumnSize();
+    }
+
     @Nonnull
     @Override
     public RecordMetaDataProto.Split toProto() throws SerializationException {