+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBStoreTimer.java
@@ -21,6 +21,7 @@
 package com.apple.foundationdb.record.provider.foundationdb;
 
 import com.apple.foundationdb.annotation.API;
+import com.apple.foundationdb.record.ScanProperties;
 import com.apple.foundationdb.record.provider.common.RecordSerializer;
 import com.apple.foundationdb.record.provider.common.StoreTimer;
 import com.apple.foundationdb.record.provider.foundationdb.keyspace.ExtendedDirectoryLayer;
@@ -446,6 +447,12 @@ public class FDBStoreTimer extends StoreTimer {
         TIME_WINDOW_LEADERBOARD_OVERLAPPING_CHANGED("number of leaderboard conditional rebuilds", false),
         /** The number of times that an index entry does not point to a valid record. */
         BAD_INDEX_ENTRY("number of occurrences of bad index entries", false),
+        /** The number of record keys repaired by {@link FDBRecordStore#repairRecordKeys(byte[], ScanProperties)}. */
+        REPAIR_RECORD_KEY("repair record key", false),
+        /** The number of record keys with an invalid split suffix found by {@link FDBRecordStore#repairRecordKeys(byte[], ScanProperties)}. */
+        INVALID_SPLIT_SUFFIX("invalid split suffix", false),
+        /** The number of record keys with an incorrect length found by {@link FDBRecordStore#repairRecordKeys(byte[], ScanProperties)}. */
+        INVALID_KEY_LENGTH("invalid record key", false),
         ;
 
         private final String title;