+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/metadata/KeyExpressionTest.java
@@ -911,6 +911,11 @@ public class KeyExpressionTest {
         public int getColumnSize() {
             return 1;
         }
+
+        @Override
+        public boolean isFixedColumnSize() {
+            return true;
+        }
     }
 
     /**
@@ -956,6 +961,11 @@ public class KeyExpressionTest {
         @Override
         public int getColumnSize() {
             return 1;
+        }
+
+        @Override
+        public boolean isFixedColumnSize() {
+            return true;
         }
     }
 }