+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/FunctionKeyIndexTest.java
@@ -409,5 +409,10 @@ public class FunctionKeyIndexTest extends FDBRecordStoreTestBase {
         public int getColumnSize() {
             return 3;
         }
+
+        @Override
+        public boolean isFixedColumnSize() {
+            return true;
+        }
     }
 }