+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/VersionIndexTest.java
@@ -206,6 +206,11 @@ public class VersionIndexTest extends FDBTestBase {
         }
 
         @Override
+        public boolean isFixedColumnSize() {
+            return true;
+        }
+
+        @Override
         public int versionColumns() {
             return 1;
         }