+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/query/QueryToKeyMatcherTest.java
@@ -687,5 +687,10 @@ public class QueryToKeyMatcherTest {
         public int getColumnSize() {
             return getArguments().getColumnSize();
         }
+
+        @Override
+        public boolean isFixedColumnSize() {
+            return getArguments().isFixedColumnSize();
+        }
     }
 }